Updated the update script to use family email address.
This should finish resolving issue #25.

There was an email address that is associated with the family, and in some cases this was present but the head of household didn't have an email set. So in those cases we now use the email address for the family for the head of household.